Update comparison functions
