+++ composer.json
@@ -26,7 +26,8 @@
             "src/Functional/Average.php",
             "src/Functional/Capture.php",
             "src/Functional/ConstFunction.php",
-            "src/Functional/Comparator.php",
+            "src/Functional/CompareOn.php",
+            "src/Functional/CompareObjectHashOn.php",
             "src/Functional/Compose.php",
             "src/Functional/Contains.php",
             "src/Functional/Difference.php",
@@ -57,7 +58,6 @@
             "src/Functional/Memoize.php",
             "src/Functional/Minimum.php",
             "src/Functional/None.php",
-            "src/Functional/ObjectHashComparator.php",
             "src/Functional/PartialAny.php",
             "src/Functional/PartialLeft.php",
             "src/Functional/PartialMethod.php",
